move the building of charts to dockerfile
The chart yamls have been moved out of ./manifests, converted to heredocs
in compilation scripts, and placed in ./charts. After downloading the
upstream chart bundles in the Dockerfile via ADD, allowing for caching,
the scripts are executed individually which will compile their
respective charts for the rke2-runtime image.

Longer term it would be nice if there was a `helmchart` cli from
helm-controller or possibly a helm plugin for compiling these.