installer: refactor install script (#220)
Make the install script simpler by leveraging a self-extracting tarball
built with makeself. This adds a new release artifact,
rke2-installer.${GOOS}-${GOARCH}.run, that is a self-extracting version
of the tarball artifact. The install, killall, and uninstall logic has
been moved into this artifact which ends up on hosts, by default, in the
/usr/local/share/rke2/scripts directory.

The install.sh at the root of the repository has been simplified. It
will attempt to detect if yum is available and install via that method.
Lacking yum, it will download and invoke the self-extracting tarball.