Add tags to prevent crashes when statically linking kubernetes binaries
This fixes MOST things, we still have a few runtime links to address but
this should at least prevent kubelet from segfaulting on startup.

Also update k3s to master, as the last update apparently didn't get us
to the rev we needed to be at to make the agent tunnel controller work
with the RBAC changes.

Signed-off-by: Brad Davidson <brad.davidson@rancher.com>