Ensure profile flag is respected
Problem: the --profile was not being respecting, causing things like the
etcd pod not being ran as the etcd user. This is because we were
shadowing a variable declation.

Fixing this also exposed that a CIS related flag,
`--protect-kernel-defaults` was improperly set.

Solution: Removing the shadowed declaration of the cisMode variable. The
global variable of the same name that is properly assigned a value
earlier on in the logic will now be used.

To fix protect-kernel-defaults, remove superfluous `--`.

Signed-off-by: Craig Jellick <craig@rancher.com>