Restore disabling built-in k3s components
This is a partial revert of #631 - we still need to --disable the k3s
built-in components to disable their controllers.

This isn't a straight revert of the change since the previous behavior
was only kind of accidentally working - the StringSlice ended up
looking like the following:

[]string{"rke2-coredns", "coredns, servicelb, traefik, local-storage, metrics-server"}

This is fine because k3s accesses the flag as a string, and urfave hides
the odd intermediate form by just joining all the args with a comma...
but if k3s ever tries to use this as an actual string slice it will not
work correctly.

Signed-off-by: Brad Davidson <brad.davidson@rancher.com>