Retag airgap tarball content to appear to come from system-default-registry
If using system-default-registry, all the manifests are updated to
specify that the image is sourced from this registry, but the images in
the tarball still come from docker.io.

Fix this by retagging all the images in the tarball to also appear to
come from the default registry. K3s actually does the work, we just have
to ask it to do so.

Signed-off-by: Brad Davidson <brad.davidson@rancher.com>