Use wharfie+k3s shared code for bootstrap
Moves the runtime image bootstrap to after config has been pulled from
the server, so that we can promote --system-default-registry to a
server var, like --cluster-dns and other things that should be set the
same on all nodes.

With this change we no longer drop manifests for packaged components on
agents, which never made use of them anyway.