kube-proxy deployed twice and without needed permissions
Kube-proxy was getting deployed twice on a fresh install. This issue was caused by the deploy controller executing before the kube-proxy startup hook was finished. This caused the chart to not get disabled before it was executed.

We also noticed that the kube-proxy pod didn't have the required permissions to the node resource.

Signed-off-by: Jamie Phillips <jamie.phillips@suse.com>