Bump vsphere chart versions
The new versions of these charts have a control plane selector that can
handle both the "controlplane" and "control-plane" label. Previously the
chart was only looking for "controlplane" and rke2 only has the
"control-plane" label.

We need to do the raw chart versions because these charts are not
officially published yet in charts.rancher.io. This problem has
highlighted a bigger process problem we have around these charts, which
we'll need to fix later. Pulling in directly from the commit has is a
bit of a band aid.

Signed-off-by: Craig Jellick <craig@rancher.com>