Get node name from metadata if AWS cloud provider is enabled (#2163) (#2166)
The AWS cloud provider expects the node name of a node to have a
specific form. If the user sets the hostname of the of the underlying
machine to a custom name, then the node will fail to bootstrap because
the node will not have the expected name.

With this change, the node name is fetched from the metadata endpoint to
ensure it has the proper value. If this fails, then it will fallback to
the hostname command.

Signed-off-by: Donnie Adams <donnie.adams@suse.com>

Co-authored-by: Donnie Adams <donnie.adams@suse.com>