Package commit as dev rpm's
Problem: RKE2 Dev and QA teams need a way to reproduce the installation
process on OS's with rpm support, without requiring to cut a new RKE2
release.

Solution: Create ephemeral unsigned rpms within the existing CI pipeline
using the same rpm spec and build scripts in rancher/rke2-packaging. The
resulting artifacts will be published on a remote repository with a short
TTL.

Caveat: These rpms are built on an Alpine environment, which uses
a slightly different FHS, requiring to take the required rpm macros from
a RHEL distribution and manually override a couple of paths.

Signed-off-by: Jonnatan Jossemar Cordero <jonnatan.cordero@suse.com>