Fix static pod UID generation and cleanup
Partial revert of 53d8510c81ff856fd5779e1b7a01f47dca46e233

This reverts static pod UID generation to use the same logic as the
kubelet, in favor of manually removing stale static pods for etcd and
the apiserver. The fixed pod UID was an attempt to trick the kubelet
into syncing static pod containers even when the apiserver was
unavailable, but it had the downside of preventing it from ever updating
the mirror pods. This was confusing for end users who expect the pods
visible via the apiserver to reflect what is actually running on the
nodes.

Signed-off-by: Brad Davidson <brad.davidson@rancher.com>