Add CLI flag for ingress controllers
* Add new --ingress-controller CLI flag
* Refactor --ingress-controller and --cni flags to use common helper for
  disabling all unused charts
* Wire first ingress controller name into global.systemDefaultIngressClass
  chart variable

NOTE: Support for the traefik chart is NOT being backported to older
minors. The flag to select an alternative ingress controller is
hidden and only supports ingress-nginx. The rest of the changes are
being backported to keep the older branches consistent with master and
allow newer releases of the ingress-nginx chart to function properly on
older minors.

Signed-off-by: Brad Davidson <brad.davidson@rancher.com>
(cherry picked from commit 76d4fb3bf430e6cee4435b02b20020907c27e2c3)
Signed-off-by: Brad Davidson <brad.davidson@rancher.com>