Add traefik airgap image tarball
Note that we are not providing a core image without ingress-nginx;
anyone who wants to use traefik instead of ingress-nginx will need to
load the core tarball (with the ingress-nginx images) in addition to the
traefik tarball. We will continue to ship ingress-nginx in the core
tarball until it is no longer the default.

Signed-off-by: Brad Davidson <brad.davidson@rancher.com>