Gate apiserver readiness on static pod sync
With the changes to the executor API that were pulled through in #7933,
apiserver readiness is now checked much faster. Unfortunately this
introduced a race condition, where startup hooks that should have waited
until after the etcd and apiserver static pods had synced are now
unblocking too early.

Fix this by having the static pod executor handle waiting for pod sync,
instead of handling this in a startup hook. This had been noted as a
TODO, and should have been done as part of the executor API changes.

Signed-off-by: Brad Davidson <brad.davidson@rancher.com>